# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "PPPK Examination Management System API with randomized questions",
        "title": "PPPKJson Exam API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/exam/{userID}": {
            "get": {
                "description": "Creates a new exam session with 20 random questions (5 per category) or returns existing active session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exam"
                ],
                "summary": "Create or get exam session",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"1234\"",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam session ready",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ExamSessionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create exam session",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    }
                }
            }
        },
        "/exam/{userID}/answer": {
            "post": {
                "description": "Submits user's answer for a specific question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exam"
                ],
                "summary": "Submit answer",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"1234\"",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Answer submission",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SubmitAnswerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Answer submitted successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Exam session not found",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to submit answer",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    }
                }
            }
        },
        "/exam/{userID}/complete": {
            "post": {
                "description": "Completes the exam session and calculates final results",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exam"
                ],
                "summary": "Complete exam",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"1234\"",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam completed successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Exam session not found",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to complete exam",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    }
                }
            }
        },
        "/exam/{userID}/results": {
            "get": {
                "description": "Retrieves detailed exam results including summary and category breakdown",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exam"
                ],
                "summary": "Get exam results",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"1234\"",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam results retrieved",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ExamResultsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Exam results not found",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    }
                }
            }
        },
        "/exam/{userID}/start": {
            "post": {
                "description": "Starts the exam timer and changes status to IN_PROGRESS",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exam"
                ],
                "summary": "Start exam",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"1234\"",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam started successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Exam session not found",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to start exam",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Get the health status of the application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": "Error message"
                },
                "message": {
                    "type": "string",
                    "example": "Success message"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.CategoryStatsResponse": {
            "type": "object",
            "properties": {
                "answered_count": {
                    "type": "integer",
                    "example": 0
                },
                "category": {
                    "type": "string",
                    "example": "MANAJERIAL"
                },
                "total_questions": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "dto.ExamResultResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "MANAJERIAL"
                },
                "exam_session_id": {
                    "type": "integer",
                    "example": 1
                },
                "grade": {
                    "type": "string",
                    "example": "B"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_passed": {
                    "type": "boolean",
                    "example": true
                },
                "max_score": {
                    "type": "integer",
                    "example": 20
                },
                "percentage": {
                    "type": "number",
                    "example": 80
                },
                "total_answered": {
                    "type": "integer",
                    "example": 5
                },
                "total_questions": {
                    "type": "integer",
                    "example": 5
                },
                "total_score": {
                    "type": "integer",
                    "example": 16
                }
            }
        },
        "dto.ExamResultsResponse": {
            "type": "object",
            "properties": {
                "results_by_category": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ExamResultResponse"
                    }
                },
                "summary": {
                    "$ref": "#/definitions/dto.ExamSummaryResponse"
                }
            }
        },
        "dto.ExamSessionResponse": {
            "type": "object",
            "properties": {
                "category_stats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CategoryStatsResponse"
                    }
                },
                "duration": {
                    "type": "integer",
                    "example": 120
                },
                "expires_at": {
                    "type": "string",
                    "example": "2026-01-28T12:00:00Z"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.QuestionResponse"
                    }
                },
                "session_code": {
                    "type": "string",
                    "example": "EXAM_1234_1643356800"
                },
                "session_id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "NOT_STARTED",
                        "IN_PROGRESS",
                        "COMPLETED",
                        "EXPIRED"
                    ],
                    "example": "NOT_STARTED"
                },
                "user_id": {
                    "type": "string",
                    "example": "1234"
                }
            }
        },
        "dto.ExamSummaryResponse": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string",
                    "example": "2026-01-28T11:30:00Z"
                },
                "exam_session_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_passed": {
                    "type": "boolean",
                    "example": true
                },
                "max_score": {
                    "type": "integer",
                    "example": 80
                },
                "overall_grade": {
                    "type": "string",
                    "example": "B"
                },
                "overall_percentage": {
                    "type": "number",
                    "example": 81.25
                },
                "total_answered": {
                    "type": "integer",
                    "example": 18
                },
                "total_questions": {
                    "type": "integer",
                    "example": 20
                },
                "total_score": {
                    "type": "integer",
                    "example": 65
                },
                "user_id": {
                    "type": "string",
                    "example": "1234"
                }
            }
        },
        "dto.QuestionOptionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 59
                },
                "option_text": {
                    "type": "string",
                    "example": "Dalam hati tidak menyetujui hal tersebut"
                }
            }
        },
        "dto.QuestionResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "enum": [
                        "MANAJERIAL",
                        "SOSIAL KULTURAL",
                        "TEKNIS",
                        "WAWANCARA"
                    ],
                    "example": "MANAJERIAL"
                },
                "exam_question_id": {
                    "type": "integer",
                    "example": 1
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.QuestionOptionResponse"
                    }
                },
                "order_number": {
                    "type": "integer",
                    "example": 1
                },
                "question_id": {
                    "type": "integer",
                    "example": 15
                },
                "question_text": {
                    "type": "string",
                    "example": "Atasan Anda melakukan rekayasa laporan..."
                }
            }
        },
        "dto.SubmitAnswerRequest": {
            "type": "object",
            "required": [
                "exam_question_id",
                "question_option_id"
            ],
            "properties": {
                "exam_question_id": {
                    "type": "integer",
                    "example": 1
                },
                "question_option_id": {
                    "type": "integer",
                    "example": 59
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}